Run generator: moved to a separate file and added FixedObsGenerator support.
